+++ internal/domain/entity/user.go
@@ -6,7 +6,7 @@ type User struct {
 	ID           uuid.UUID
 	Login        string
 	PasswordHash string
-	Current      int
+	Current      float32
 }
 
 func NewUser(login, passwordHash string) *User {
